Improve consistency with other docs
We are missing basic branding on our cookbook docs, which make the
cookbooks look very unofficial. This PR adds favicon and Arrow logo. It
also matches the body font selection to match our main documentation site.

In addition, links are added to the cookbook Github and to each languages'
respective user guide and API reference.