Improve consistency with other docs (#300)
We are missing basic branding on our cookbook docs, which make the
cookbooks look very unofficial. This PR adds favicon and Arrow logo. It
also matches the body font selection to match our main documentation
site. This isn't meant to be a full overhaul of the style; I just wanted
to make the minimal changes so that a visitor would feel that this is
part of the official Apache Arrow documentation.

In addition, links are added to the cookbook Github and to each
languages' respective user guide and API reference.