MINOR: Fix branch protection in .asf.yaml (#379)
The `.asf.yaml` has been processed by a new, stricter system but we
didn't have any changes since it was activated, so the issue with the
dummy key didn't show up. I was informed through an [email
notification](https://lists.apache.org/thread/v75s5wfnslmsz6soo5pstplb8205vkj4)
of this issue.

I have not opened this as a fork PR to enable the notification via email
in case there is still an issue.