[JAVA][CI] Adds a recipe for Flight SQL
A very basic database server is implemented that uses the DataFusion
query engine for answering SQL queries on single-file, write-once tables
stored in the Arrow IPC format.

Although this does not implement all the features and possibilities of
FlightSqlProducer, it show cases the two main categories of concepts
through the inclusion of tables (like catalog, database schema, sqlinfo,
etc) and statement (like prepared statement and substrait statement).

Some updates to the building of the Java Cookbook:
Using the release property instead of the compiler source and target
values which is deprecated.